requirements.txt
Django==4.2.7
django-simple-history==3.4.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sentry-sdk==1.39.0
openpyxl==3.1.2
pandas==2.1.4
flake8==6.1.0
black==23.12.1
bandit==1.7.6